(maint) Improve overall extendability
This commit adds the `common` folder, which, for now, stores
the `extension.py` file. This file is responsible for housing
the base class for all extensions in the app and the format
enforcement of extensions. Extensions will be recognized only
if they extend the `Extension` class and they must have a
setup method with the exact same signature. Alignment of
extensions with these changes are also done.

This also enables the config file to be changed via argument
while maintaining intended overall functionality (full arg
help, load config before all extensions, make config available
to extensions etc).

Other small improvements and rewriting of code are also done.